Линтер зелёный, проверка готовности проходит
ruff нашёл восемь замечаний в src и tests — их запускает CI, а я за ночь
ни разу не прогонял. Шесть исправлены автоматически, два руками:
однобуквенное имя переменной в подмене и лишний int вокруг round.

vantage doctor: блокирующих проблем нет, семь параметров экономики
подтверждены источником, четыре — инженерная оценка и названы поимённо.

Логи прогонов удалены из рабочей копии.

673 проверки.

+++ src/vantage/risk.py
@@ -480,7 +480,7 @@ def _classify_top(risk: pd.Series, *, quantiles: int = 4) -> pd.Series:
     if live.empty:
         return out
 
-    take = max(1, int(round(len(risk) * PUBLIC_TOP_SHARE)))
+    take = max(1, round(len(risk) * PUBLIC_TOP_SHARE))
     # method="first" вместо среднего: при одинаковом риске у сотни ячеек
     # средний ранг даёт им общее место, и отбор берёт либо все сто, либо
     # ни одной.
